Fix a regression on MapBasedTreeState
MapBasedTreeState was returning itself when the next transform could
use a cached TreeState. The problem is that this class also contains
notified mutable field. So the effect was that once we got a cached
version and used once, we would never go back to the FS based version.

Added also some basic tests on TreeState (In addition to what we
have in Sequence).

Change-Id: If81365368d35a7d2218299b8cacf8d6f1c162623