Multiple remote refs with same name for local git repos
This change allows to create references in the local repository for
git.destination even when there are multiple remote references that
match the push reference (For example 'master' matches
'refs/heads/master' and 'refs/for/master'). The usage of this name
is purely cosmetic, so that the user gets a git client with a
proper reference/branch name.

I also removed ls-remote dependency since, as said before, this
is purely cosmetic. We use remote.origin.push for the push.

BUG=36879436

Change-Id: Ie510509600b1f016e8ba0702bb35c3d4c6cf5108