Make label group manipulation in message more robust
This change make add/replace/remove operations on labels
more robust. Instead of using regex on arbitrary parts
of the message we create an object that contains the distinct
parts of the message and we use that object for manipulating
the message.

Also changed git destinations to use this new mechanism for
adding the labels. This fixes a bug where on git.destination,
if the user added a label like "Change-Id", we wouldn't include
it in the last paragraph. This made Gerrit not recognize it.

BUG=36895806

Change-Id: I14a0cc887a84c53ca1ee5fcd7e3d70dd55dcfc93