Use workflowId to compute ChangeId
By basing the ChangeId solely on data known a-priori, we
can query Gerrit for the specific migration without using
weak criteria like origin change (which might not be unique).

To avoid running into issues with repeated exports, we
keep an attempt counter in the hash.

The default implementation rehashes the changeID with the time
to guarantee unique IDs (never finds a change).
BUG=37512895

Change-Id: I6be59a387a4364eab3174187f05d966319395ea1