Generic dry-run mode for workflows
Currently we have a dry mode for git destination but this should be
a generic mechanism (We might keep the git.destination flag since it
is very explicit on what it does).

The advantage of this method is that it fails if the destination
doesn't support dryrun.

I also plan to use this mechanism for a different problem: Being able
to validate if detected --last-ref change is what Copybara generated
or there are manual changes done in the destination change or later
change. This will work by validating that a migration of --last-rev
revision on dryrun mode generates an empty change.

PiperOrigin-RevId: 154183721
Change-Id: I5abfd7ca2e99f8342eacef8f42c807b0ee2fba07