Implement group identity
This is the initial implementation to have support for a group
identifier for a migration. For example a Github pull request or
a gerrit topic.

We need this because:

  - We need to be able to group changes in one entity (PR for example)

  - We need to be able to query that entity for getting the latest
    revision migrated so that we don't accidentally migrate a
    previous revision. (Will come in next CLs).

  - This is specially important for system that can give you stale
    data.

This implementation should be a no-op, since it is designed so that
in order to work, both origin and destination need to support and
be able to identify the group identity.

PiperOrigin-RevId: 156101550
Change-Id: I2d6f991ead593c19ed69f2cbd308c7bf94702ff2