Record Gerrit patchset number in Git Origin-RevId
The format used is:
  SHA-1 (SPACE ARBITRARY_STRING)?

This allows regular git.origin or other git origins to
parse SHA-1. But allow the specialized versions to
record more data.

This is needed so that we can ensure that when we
revert a commit in gerrit we migrate that again.

Note that this is not really used for now since we only
compare with the baseline.

PiperOrigin-RevId: 156822652
Change-Id: Ic685b1c4f9634fcf459cfcb1f55b44a11b8a8dca