Refactoring the reuse of output dirs.
The previous implementation had the side effect of deleting
the output dirs in the middle of the run if they existed.
Deleting the output dirs at the beginning is more stable.
I am creating a new subdirectory output in which the
run specific directories will go. Right now they are all
of them, but in future we will move directories that we
want to survive between runs directly in the root path.

PiperOrigin-RevId: 160461586
Change-Id: I6013cadbb2b008a3ae7c94f191a35e535b76cd77