Merge Destination Reader and Writer
It doesn't make sense to have them separate as newWriter/newReader
acts as a factory for the workflow and this prevents to keep state
between the two.

It also makes more evident missing implementations of Reader.

Last point: It makes the Workflow helper less confusing, as we had
two readers, one for destination and one for origin.

PiperOrigin-RevId: 161569209
Change-Id: I31626f25944e2e4ac54bc56c0ca2435100eace67