git credential fill tool for non-interactive
This change adds a utility class for reading credentials using
'git credential fill' command. This allows to plug copybara
with git helpers like store, cache or wallet based ones.

The plan is to configure the internal local repos we create
to use the helper "store --path ..." so that it tries to use
the store helper. But this allows that if the store doesn't
exist or is empty to delegate to other helpers configured in
$HOME/.gitconfig.

This doesn't support asking for passwords since:

  - It is really hard, we need to build a binary and connect it
    with sockets to the current one.

  - Even if we had it we would need to use that same program with
    all our git invocations to be worth it (Otherwise git would
    ask for password anyway).

BUG=33387564
Change-Id: I66ba77c7c78dde076c434cdf2bbc15a190d16e10