Support for repeated labels
Up to now we were only keeping the last label found in the message.
This is limiting in several circumstances (For example if it appears
multiple times). This is specially true in git, since labels like
Signed-off-by are meant to appear multiple times, and the order has
meaning (Is a chain of signed offs).

This change adds the internal wiring for supporing multiple
appearances, adds the support for git.origin (In fact the code is
simpler) and exposes the the multiple values as a new method in
Skylark.

find_all_labels returns all the labels found no matter the source.
Some users might only want to look in changes (but they can use
change.labels_all_values) or in the message (Not supported for now).
If there is a common usage pattern we might make find_all_labels more
flexible in the future (or add a new method).

BUG=64814355
Change-Id: Ibd527454befac3666db8f28ec3d8008d4284818d