Create github pull requests in github pr destination
The feature can be disabled if needed. We check first if there
are any open PR for the identity branch name. If not, we create
one.

For now we don't recreate if the baseline is different.

We use the current change first line and message for title and
body. This is not ideal but it works well for 1 change import.
We'll probably add something in the future to make all the changes
visible or defer the PR creation to the last import so that it
can access all the previous changes.

Change-Id: Ie4e571968897a9d946c28e20a989d201f57ceb58