RecordsProcesscalldestination groupId != pending
Group identity can be not-null for non-pending changes in the origin.
The original code made the assumption that the change should be
pending in the origin if it the destination groupId != null. But that
is not true. For example in a workflow for git submitted -> GitHub PR
the group identity would be something like "master" or "origin_branch"
in the destination but the change is not pending in the origin.

Change-Id: Ia11f7519397ed31e5b8f97073deb2192f5d430ac