CapturingConfigFile.relativeToRoot should delegate
This class was relying on the default behavior instead of using the
wrapper behavior. Not returing the relative path as an effect.

This was breaking the identity.

I made the method abstract to force the overwrite/implementation.

This is expected to break identity compatibility with old code. But
it is already broken unless absolute path are the same.

BUG=65539809
PiperOrigin-RevId: 168389946
Change-Id: I4d0a9d67fdba9ec8e497eb9083ffbe7ffa8b0fc4