Include files in git.origin
This change returns the files for git origin changes. Allowing
the workflows to filter out non-affecting change.

As part of this change, LogCmd was changed to allow getting the
files for a merge.

We still include the message for all the changes in a branch
when using includeBranchCommitLogs. But that requires a bit more
thinking.

BUG=63576333

Change-Id: Ib950beb1742d3956937795738d8ac259cbfe5904