Make CHANGE_REQUEST work with metadata.squash_notes
CHANGE_REQUEST doesn't include all the changes between the baseline
and the HEAD. While this is understandable if we receive a manual
baseline via the flag, we want to include all the imported changes
in the normal scenario.

This allows to use metadata.squash_notes for multi-commit
CHANGE_REQUEST imports.

The main use case is to use with git.github_pr_origin.

BUG=66954620

Change-Id: I3250b02bcf95331365910836b96cd43ee99f399e