Remove only implementation of Changes
Changes now only has one implementation (We removed the lazy one,
time ago). So it is much simpler to use Changes directly.

We had some tests that throw UOException on some method calls, but
this is a minor thing and if they care they should check the result
of the getCurrent/getMigrated methods (We return empty).

Change-Id: Ifb440071903ae898f78924de1d064ed5beda2039