Allow pagination in git log
This adds support for the --skip flag. In combination with -n x, it
allows us to do pagination.

Wrote a test that checks that it gives the same result even when
using -m --name-only (that contains multiple entries per merge).

Added also a test that validates that the first entry for a merge
is always the current branch merge (and contains the difference
of files with the rest of the branches).

BUG=69804529,34449319

Change-Id: I9d665a21499eb01e3fb8ab03ac1001950cbe877f