Limit available to migrate changes printed
If a lot of changes are available (say, the migration
was never run) right now we print all of them to the
console. This could be thousands.

Introduce a limit in the console, configurable by a flag
that defaults to 25 (fits well in the terminal) and log
also how many were not printed.

This does not affect structured output. All the
changes are populated there.

PiperOrigin-RevId: 178442617
Change-Id: Ib93147db9c66737617a5a2c81f1f40e3b7257e85