Expose several core info as labels
This change exposes last-rev, current revision, message and message
title as labels so that it can be used by our template message
manipulators or Skylark transformations.

For now we compute the map of labels everytime, since it is almost
free. If this turns to be a problem we can precompute it.

With this we can construct messages like:

  Import diff https://github.com/foo/bar/compare/abcd..bcde
or
  Import of message_title

BUG=71714950

Change-Id: I87e52c4fd90d86bb89de317db3a4360b827f7f12