Expose a couple of new Skylark transform functions
The new functions are:

  now_as_string: I could split it in two but for now it is good enough.
  I was going to allow formatting any timestamp, but then I found
  (again!) that Skylark doesn't support Long.

  read_path: Read a path from the checkout dir

  write_path: Write an arbitrary string to a path in the checkout dir.

Yes, this opens the door to a lot of awful transformations, but OTOH
it is in the Skylark transformation code. Also, nothing that they
cannot do with core.copy + core.replace (Except less hacky). These
kind of tranformations are not mean to be reversable.

Example usage:

  def test(ctx):
      path = ctx.new_path('folder/file.txt')
      ctx.read_path(path)
      ctx.write_path(path, ctx.read_path(path) + ctx.now_as_string())

FIX=72655061

Change-Id: Idb2a5f0bce9f23309d853b36680a67ec50070306