Allow params in dynamic transformations
Dynamic transformations don't accept parameters. This makes it
difficult to use them inside libraries. Since they cannot be
customized.

This change allows to explicitly create the dynamic transformation
in skylark via 'core.dynamic_transformation' and pass customized
parameter. The parameters are available under ctx.params dict. Note
that this change is backward compatible with the existing dynamic
transformations, that are easier to use for single usages.

As part of this change I also extended @Example to allow testing
snippets that are not expressions. We might revisit this in the
future to make it even more complete (for example for core.workflow
examples). But for now it is good enough.

Change-Id: Idceb13fb21c7cc15822ab7b4444318b30d844e45