github_pr_origin migrates only open PR by default
Currently we migrate all PR no matter what state they are. This
change allows to select what kind of state the PR has to be in
order to migrate.

This changes the behavior since the old behavior was 'ALL' and
now the default is 'OPEN'. But that should be fine, since most
if not all of our users are interested in open PR.

FIX=73356561
Change-Id: Icd887a227697c89a93abfedfd5b8198c1a475945