Implement CHANGE_REQUEST_FROM_SOT workflow mode
Currently, if we want a sane diff, we can only import pending changes
to the SoT. If we try to use CHANGE_REQUEST in the other direction, it
fails because we cannot find the label in the origin. If we use
instead SQUASH, we end up with a bigger diff than the change itself.

This new mode allows, if the origin pending change baseline is already
in the destination, to use a CHANGE_REQUEST like import to the
destination.

Two implementations are provided: Gerrit and GitHub PR. It doesn't
make sense for the other cases (For now).

Note that in case we cannot find the baseline in the destination,
we fail instead of doing SQUASH. We might revisit this in the
future.

BUG=69857795
PiperOrigin-RevId: 187229253
Change-Id: Ic9f4890b1953196e3d9af57d96c5b92e7ec32c1a