Use destination_files glob for finding last rev
Before this change we only used the roots (and label)
to find the last change, but that doesn't work if the
excluded files are a complex glob inside the included
one. For example:

destination_files = glob([foo/**], exclude = [foo/bar/**])

This allows users to have different migrations with
the same Git repo as a destination, having mutually
exclusive destination_files.

FIXED=74118492
PiperOrigin-RevId: 187655805
Change-Id: I8c80ea3925302346d648134d5a9802db9f5d01f4