Allow constant title and body for PR destination
Currently we set the title/body to the one in the first commit.
This works well for single commit exports (Gerrit origin for
example), but bad when used with iterative, since the first
commit is meaningless for the overall export.

We might make it more flexible in the future but this is good
enough for now.

FIX=74151156

Change-Id: Ib23b54c8b51ba0bde34f2c96b7cbe8c6f629ae88