Retries for CHANGE_REQUEST_FROM_SOT
Currently if CHANGE_REQUEST_FROM_SOT cannot find the baseline in the
destination it directly fails. In the tool this is expected, since
the user would need to re-run the post-submit workflow.

Internally we have a service that runs workflows independently. There
is a race condition where the pending change might sync to head
before running but the postsubmit workflow might still be pending
or running.

The ideal solution for this would be to have dependant workflows,
since this would ensure the state of the postsubmit is known before
running the presubmit one. But this introduces some technical
complexities for our service.

Instead, for now, we introduce a flag to do some retries in the
tool and we also return REPOSITORY_ERROR so that our service
can run a retry.

As part of this I've introduced two things that I am not very
proud:

  - Wrap InterruptedException in RepoException: We should propagate
    interrupted instead. LocalParallelizer has the same problem.
    But not in this change.

  - ValidationException can be retried. I guess I could just throw
    RepoException. But the thing is that we know this is a user error
    if we cannot find the baseseline after a while. We might
    expose a new exit code for retryable user errors or just change
    this to RepoException in the future.

FIX=74263970
Change-Id: I0fadb648784c06e15ee08d774136d23b00c6b196