Make reader.changes() return a reason for no results
The current approach tries to do some guessing at the workflow level
based on heuristics (like changes(to, from) , etc.). This is
wrong in several cases since we don't have access to the real
history and more expensive ( not much more, but git merge-base is
more efficient than parsing all the log for git log).

FIX=77150563
PiperOrigin-RevId: 190819216
Change-Id: Icd689c8057f64143986bb4fbf6a9360a94c367db